% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfOT.R
\name{acfOT}
\alias{acfOT}
\title{ACF of a dichotomized process}
\usage{
acfOT(PD, ACSn, fast = 0)
}
\arguments{
\item{PD}{A scalar (0<=PD<1), denoting probability level of the threshold at which dichotomization is applied.}

\item{ACSn}{A vector, with the equivelant (i.e., Gaussian) autocorrelation structure.}

\item{fast}{Boolean, TRUE or FALSE, indicating whether to use or not an approximation of the bivariate Gaussian copula (default valueis FALSE)}
}
\value{
A vector (with length equal to ACSn) with the ACF of the dichotomized process.
}
\description{
ACF of a dichotomized process
}
\examples{
 # Define an AR(1) ACF, with rho=0.6, up to lag 100.
 ACS=0.6^(0:100)

 # Estimate the over-threshold ACF using the "fast", and the "slow" method.
 acfOTfast=acfOT(PD = 0.7, ACSn = ACS, fast = 1)
 acfOT=acfOT(PD = 0.7, ACSn = ACS, fast = 0)

 # Compare the fast", and the "slow" method.
 plot(acfOTfast, acfOT)
 abline(0,1,col='red')

}
