% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmaxxer.R
\name{blockmaxxer}
\alias{blockmaxxer}
\title{Function to exctract block-maxima}
\usage{
blockmaxxer(x, k, n, fun = max)
}
\arguments{
\item{x}{a vector with the time series (not xts)}

\item{k}{length of block}

\item{n}{number of blocks}

\item{fun}{function to apply (default is max)}
}
\value{

}
\description{
Function to exctract block-maxima
}
