% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian2binary.R
\name{Gaussian2binary}
\alias{Gaussian2binary}
\title{Fast estimation of equivelant correlation coefficients for Bernoulli processes}
\usage{
Gaussian2binary(P, ACSn)
}
\arguments{
\item{P}{parameter of Bernoulli process}

\item{ACSn}{A vector, with the equivelant (i.e., Gaussian) autocorrelation structure.}
}
\value{
vector of equivelant correlation coefficients
}
\description{
Fast estimation of equivelant correlation coefficients for Bernoulli processes
}
