% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICcorr.R
\name{ICcorr}
\alias{ICcorr}
\title{Estimate the Intra-cluster (IC) correlation coefficient}
\usage{
ICcorr(ACFn = acsHurst(H = 0.6, lag = 99), kmax = 100, SV = 0)
}
\arguments{
\item{ACFn}{A vector with the autocorrelation structure (length kmax<=m)}

\item{kmax}{Maximum scale (kmax<=m)}

\item{SV}{Boolean (TRUE or default: FALSE) indicating if IC will be calculated in the intermidiate time scales}
}
\value{
The Intra-cluster (IC) correlation coefficient
}
\description{
Estimate the Intra-cluster (IC) correlation coefficient
}
\examples{

k=365
ACF=anySim::acsCAS(param = c(3, 0.1), lag = k-1)
sum(SuperGauss::toep.mult(acf = ACF, rep(1,k)) -1)/(k*(k-1))
(sum(toeplitz(ACF))-k)/(k*(k-1))
pracma::tic();ICcorr(ACFn = ACF, kmax = k, SV = 1);pracma::toc()

}
