% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pAR1.R
\name{pAR1}
\alias{pAR1}
\title{Distribution of extremes (maxima) of a stochastic process with AR(1) autocorrelation structure and any marginal distribution}
\usage{
pAR1(q, dist, params, rhoz, k)
}
\arguments{
\item{dist}{quantile function (character, e.g., qgamma)}

\item{params}{a list with distribution's CDF (as function) and parameters}

\item{rhoz}{A scalar, with the lag-1 equivelant (i.e., Gaussian) correlation coefficient}

\item{k}{duration (e.g., 365) or block length}

\item{u}{vector of probabilities}
}
\value{
vector of probabilities
}
\description{
Distribution of extremes (maxima) of a stochastic process with AR(1) autocorrelation structure and any marginal distribution
}
