% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pARp.R
\name{pARp}
\alias{pARp}
\title{Distribution of extremes (maxima) of a stochastic process with AR(p) autocorrelation structure and any marginal distribution}
\usage{
pARp(q, dist, params, ACSn, k, p, fast = 0)
}
\arguments{
\item{ACSn}{A vector, with the equivelant (i.e., Gaussian) autocorrelation structure.}

\item{k}{duration (e.g., 365) or block length}

\item{p}{AR(p) model order}

\item{fast}{Boolean, TRUE or FALSE, indicating whether to use or not an approximation of the bivariate Gaussian copula (default valueis FALSE)}

\item{u}{vector of probabilities}
}
\value{
vector of probabilities
}
\description{
Distribution of extremes (maxima) of a stochastic process with AR(p) autocorrelation structure and any marginal distribution
}
