% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbbextremes.R
\name{pbbextremes}
\alias{pbbextremes}
\title{Distribution of extremes (maxima) of a stochastic process with any autocorrelation structure (ACS)}
\usage{
pbbextremes(q, dist, params, ACSn = ARTApar$ACSn, fast = 0)
}
\arguments{
\item{ACSn}{A vector, with the equivelant (i.e., Gaussian) autocorrelation structure.}

\item{fast}{Boolean, TRUE or FALSE, indicating whether to use or not an approximation of the bivariate Gaussian copula (default valueis FALSE)}

\item{u}{vector of probabilities}
}
\value{
vector of probabilities
}
\description{
Approximation based on the beta-Binomial distribution. For more details see, Serinaldi F, Lombardo F, Kilsby CG (2020) All in order: Distribution of serially correlated order statistics with applications to hydrological extremes. Adv Water Resour 144:103686. doi: 10.1016/j.advwatres.2020.103686.
}
