% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbivnorm.R
\name{pbivnorm}
\alias{pbivnorm}
\title{Approximation of the bivariate normal CDF}
\usage{
pbivnorm(q, rho)
}
\arguments{
\item{q}{vector of quantiles}

\item{rho}{Pearson's correlation}
}
\value{
vector of probabilities
}
\description{
Approximation of the bivariate normal CDF
}
\examples{

rho=0.4
Z=MASS::mvrnorm(n = 10^3, mu = c(0, 0), Sigma = matrix(c(1,rho, rho,1), ncol = 2, byrow = T) )

mz=matrixStats::rowMaxs(Z)
z=seq(-4, 4, by=0.01)
p=pbivnorm(q = z, rho)

plot(sort(mz), ppoints(mz, a=0))
lines(sort(z), p, col='red')
lines(sort(z), pnorm(z)^2, col='blue')

rho=0.6
Z=seq(-4,4,by=0.1)

pz=rep(NA, length(Z))
for (i in 1:length(Z)) {
  pz[i]=mvtnorm::pmvnorm(mean = c(0,0), upper = Z[i], corr = matrix(c(1,rho, rho,1),  ncol = 2, byrow = T))[1]
}


z=seq(-4, 4, by=0.01)
p=pbivnorm(q = z, rho)

plot(Z, 1-pz, log='y')
lines(sort(z), 1-p, col='red')
points(sort(z), 1-p, col='red', pch=19, cex=0.2)
lines(sort(z), 1-pnorm(z)^2, col='blue')

plot(pz, pbivnorm(q = Z, rho)); abline(0,1, col='red')


}
