% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcopmax.R
\name{pcopmax}
\alias{pcopmax}
\title{Distribution of extremes (maxima) of a stochastic process with any autocorrelation structure (ACS) using Copulas}
\usage{
pcopmax(q, dist, params, ACSn = ARTApar$ACSn)
}
\arguments{
\item{ACSn}{A vector, with the equivelant (i.e., Gaussian) autocorrelation structure.}

\item{u}{vector of probabilities}
}
\value{
vector of probabilities
}
\description{
Distribution of extremes (maxima) of a stochastic process with any autocorrelation structure (ACS) using Copulas
}
