% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFX.R
\name{pFX}
\alias{pFX}
\title{cummulative distribution funcion (CDF)}
\usage{
pFX(q, FX, PFX)
}
\arguments{
\item{q}{vector of quantiles}

\item{FX}{CDF function (character, e.g., pgamma)}

\item{PFX}{a list with distribution's CDF (as function) and parameters}
}
\value{
vector of probabilities
}
\description{
cummulative distribution funcion (CDF)
}
