% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qFX.R
\name{qFX}
\alias{qFX}
\title{Quantile function (Inverse cumulative distribution function)}
\usage{
qFX(p = ppoints(1000, a = 0), FX, PFX)
}
\arguments{
\item{p}{vector of probabilities}

\item{FX}{quantile function (character, e.g., qgamma)}

\item{PFX}{a list with distribution's ICDF (as function) and parameters}
}
\value{
vector of quantiles
}
\description{
Quantile function (Inverse cumulative distribution function)
}
