% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piid.R
\name{qmaxiid}
\alias{qmaxiid}
\title{Quantile function (Inverse cumulative distribution funtcion) of the distribution of k-length block maxima under the assumption of temporal independence}
\usage{
qmaxiid(Rp, QFX, k, ...)
}
\arguments{
\item{Rp}{a scalar denoting the return period (in years)}

\item{QFX}{the quantile function of distribution of the parent process}

\item{k}{a scalar denoting the length of block (e.g., 365 for 1 year, in the case of a daily parent process)}

\item{...}{additional arguments (parameter) that regards the distribution of the parent process}
}
\value{
vector of return levels
}
\description{
Quantile function (Inverse cumulative distribution funtcion) of the distribution of k-length block maxima under the assumption of temporal independence
}
