% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsuperG.R
\name{simsuperG}
\alias{simsuperG}
\title{Fast simulation of stationary univariate processes with any marginal distribution and autocorrelation structure}
\usage{
simsuperG(
  n = 1,
  ACF,
  dist,
  params,
  NatafIntMethod = "GH",
  NoEval = 9,
  polydeg = 0,
  stand = 0
)
}
\arguments{
\item{n}{number of realizations}

\item{ACF}{A vector with the target autocorrelation structure (including lag-0, i.e., 1). The length of the ACF vector should be equal to that of the time series to be generated.}

\item{dist}{A string indicating the quantile function of the target marginal distribution (i.e., the ICDF).}

\item{params}{A named list with the parameters of the target distribution.}

\item{NatafIntMethod}{A string ("GH", "Int", or "MC"), indicating the intergation method, to resolve the Nataf integral. Currently only "GH" is implemented.}

\item{NoEval}{A scalar indicating (default: 9) the number of evaluation points for the integration methods.}

\item{polydeg}{A scalar indicating the order of the fitted polynomial. If polydeg=0, then another curve is fitted.}

\item{stand}{Boolean indicator (TRUE or default: FALSE) whether or not standardization is performed.}
}
\value{
A list with 5 elements. Two (2) that concern the Nataf transofrmation (i.e., nataf and acfz), and three (3) that regard the generated time series (in vector format): X: The final time series at the actual domain with the target marginal distribution and correlation structure; Z: The auxiliary Gaussian time series at the Gaussian domain and; U: The auxiliary uniform time series at the Copula domain (i.e., in [0,1]).
}
\description{
Fast simulation of stationary univariate processes with any marginal distribution and autocorrelation structure
}
